pyproject.toml
[project]
name = "sirene-api-client"
version = "0.1.0"
description = "Professional Python client for the French SIRENE API with comprehensive ETL capabilities. Query French business registry data (SIREN/SIRET), extract complete company histories, transform coordinates. Features type-safe Pydantic models, async/sync support."
authors = [
    {name = "Francesco Portigliotti", email = "francesco.portigliotti@karibu.earth"},
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "httpx>=0.27.0",
    "attrs>=23.2.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.1.1",
    "pydantic>=2.12.0",
    "pyproj>=3.6.1",
]

[project.urls]
Homepage = "https://github.com/karibu-earth/sirene-api-client"
Repository = "https://github.com/karibu-earth/sirene-api-client"
Documentation = "https://github.com/karibu-earth/sirene-api-client#readme"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "sirene_api_client"
module-root = ""
source-include = ["CHANGELOG.md", "sirene_api_client/py.typed"]

[tool.ruff]
line-length = 88  # Changed from 120 to match Black standard
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "RUF", # ruff-specific rules
    "PT",  # flake8-pytest-style
    "A",   # flake8-builtins
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["sirene_api_client"]
force-sort-within-sections = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"

[tool.ruff.lint.flake8-type-checking]
runtime-evaluated-base-classes = ["pydantic.BaseModel"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.13"  # Target Python 3.13 only
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict = true
explicit_package_bases = true
exclude = [
    "tests/*",
    "examples/*",
]

[tool.bandit]
exclude_dirs = ["tests", "examples"]
skips = ["B101"]  # Skip assert_used in test files

[tool.coverage.run]
source = ["sirene_api_client"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/examples/*",
]

[tool.coverage.report]
fail_under = 90  # TDD policy requirement
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "if TYPE_CHECKING:",
]

[dependency-groups]
analysis = [
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.9.4",
    "mypy>=1.18.2",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "seaborn>=0.13.2",
]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
    "types-python-dateutil>=2.9.0.20250822",
]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",  # HTTP mocking for httpx
    "pytest-xdist>=3.7.0",  # Parallel testing
]
lint = [
    "ruff>=0.13.3",
    "pre-commit>=4.2.0",
]
typing = [
    "mypy>=1.18.2",
    "types-python-dateutil>=2.9.0.20250822",
]
security = [
    "bandit>=1.8.5",
    "detect-secrets>=1.5.0",
    "pip-audit>=2.9.0",
]
